Renamed DaggyCore. hpp instead of h for cpp headers
