+++ src/DaggyCore/CJsonDataSourcesConvertor.cpp
@@ -21,10 +21,10 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-#include "Precompiled.h"
-#include "CJsonDataSourcesConvertor.h"
+#include "Precompiled.hpp"
+#include "CJsonDataSourcesConvertor.hpp"
 
-#include "Common.h"
+#include "Common.hpp"
 
 using namespace daggy;
 using namespace daggyconv;
