+++ src/DaggyCore/CLocalDataProvidersFabric.cpp
@@ -21,9 +21,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-#include "Precompiled.h"
-#include "CLocalDataProvidersFabric.h"
-#include "CLocalDataProvider.h"
+#include "Precompiled.hpp"
+#include "CLocalDataProvidersFabric.hpp"
+#include "CLocalDataProvider.hpp"
 
 using namespace daggy;
