+++ src/DaggyCore/CSsh2DataProvider.cpp
@@ -21,11 +21,11 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-#include "Precompiled.h"
-#include "CSsh2DataProvider.h"
+#include "Precompiled.hpp"
+#include "CSsh2DataProvider.hpp"
 
-#include "Ssh2Client.h"
-#include "Ssh2Process.h"
+#include "Ssh2Client.hpp"
+#include "Ssh2Process.hpp"
 
 using namespace daggy;
 using namespace daggyssh2;
