+++ src/DaggyCore/Common.hpp
