+++ src/DaggyCore/IDataAggregator.cpp
@@ -21,8 +21,8 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-#include "Precompiled.h"
-#include "IDataAggregator.h"
+#include "Precompiled.hpp"
+#include "IDataAggregator.hpp"
 
 using namespace daggy;
