+++ src/DaggyCore/IDataProviderFabric.cpp
@@ -21,10 +21,10 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-#include "Precompiled.h"
-#include "IDataProviderFabric.h"
+#include "Precompiled.hpp"
+#include "IDataProviderFabric.hpp"
 
-#include "Common.h"
+#include "Common.hpp"
 
 using namespace daggy;
