+++ src/DaggyCore/Result.hpp
