+++ src/DaggyCore/Ssh2Types.cpp
@@ -21,8 +21,8 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 */
-#include "Precompiled.h"
-#include "Ssh2Types.h"
+#include "Precompiled.hpp"
+#include "Ssh2Types.hpp"
 
 using namespace daggyssh2;
